separated some utils from excellib
