+++ src/pycel/excelutil.py
@@ -430,5 +430,101 @@ def uniqueify(seq):
     seen_add = seen.add
     return [ x for x in seq if x not in seen and not seen_add(x)]
 
+def is_number(s): # http://stackoverflow.com/questions/354038/how-do-i-check-if-a-string-is-a-number-float-in-python
+    try:
+        float(s)
+        return True
+    except ValueError:
+        return False
+
+def is_leap_year(year):
+    if not is_number(year):
+        raise TypeError("%s must be a number" % str(year))
+    if year <= 0:
+        raise TypeError("%s must be strictly positive" % str(year))
+
+    return year % 4 == 0 and year % 100 != 0 or year % 400 == 0
+
+def get_max_days_in_month(month, year):
+    if not is_number(year) or not is_number(month):
+        raise TypeError("All inputs must be a number")
+    if year <= 0 or month <= 0:
+        raise TypeError("All inputs must be strictly positive")
+
+    if month in (4, 6, 9, 11):
+        return 30
+    elif month == 2:
+        if is_leap_year(year):
+            return 29
+        else:
+            return 28
+    else:
+        return 31
+
+def normalize_year(y, m, d):
+    if m <= 0:
+        y -= int(abs(m) / 12 + 1)
+        m = 12 - (abs(m) % 12)
+        normalize_year(y, m, d)
+    elif m > 12:
+        y += int(m / 12)
+        m = m % 12
+
+    if d <= 0:
+        d += get_max_days_in_month(m, y)
+        m -= 1
+        y, m, d = normalize_year(y, m, d)
+
+    else:
+        if m in (4, 6, 9, 11) and d > 30:
+            m += 1
+            d -= 30
+            y, m, d = normalize_year(y, m, d)
+        elif m == 2:
+            if (is_leap_year(y)) and d > 29:
+                m += 1
+                d -= 29
+                y, m, d = normalize_year(y, m, d)
+            elif d > 28:
+                m += 1
+                d -= 28
+                y, m, d = normalize_year(y, m, d)
+        elif d > 31:
+            m += 1
+            d -= 31
+            y, m, d = normalize_year(y, m, d)
+
+    return (y, m, d)
+
+def year_from_int(nb):
+    if not is_number(nb):
+        raise TypeError("%s is not a number" % str(nb))
+
+    # origin of the Excel date system
+    current_year = 1900
+    current_month = 0
+    current_day = 0
+
+    while(nb > 0):
+        print current_year, current_month, current_day, nb
+        if not is_leap_year(current_year) and nb > 365:
+            current_year += 1
+            nb -= 365
+        elif is_leap_year(current_year) and nb > 366:
+            current_year += 1
+            nb -= 366
+        else:
+            current_month += 1
+            max_days = get_max_days_in_month(current_month, current_year)
+            
+            print 'Max days', max_days, current_month
+            if nb > max_days:
+                nb -= max_days
+            else:
+                current_day = nb
+                nb = 0
+
+    return (current_year, current_month, current_day)
+
 if __name__ == '__main__':
     pass
